新增 角色六属性(力敏体智精魅)达到客户端极限(255) 选择角色画面显示异常待解决(疑似客户端极限)

+++ trunk/src/l1j/server/server/storage/mysql/MySqlCharacterStorage.java
@@ -71,12 +71,12 @@ public class MySqlCharacterStorage implements CharacterStorage {
 			pc.setStatus(0);
 			pc.addBaseMaxMp(rs.getShort("MaxMp"));
 			pc.setCurrentMpDirect(rs.getShort("CurMp"));
-			pc.addBaseStr(rs.getByte("Str"));
-			pc.addBaseCon(rs.getByte("Con"));
-			pc.addBaseDex(rs.getByte("Dex"));
-			pc.addBaseCha(rs.getByte("Cha"));
-			pc.addBaseInt(rs.getByte("Intel"));
-			pc.addBaseWis(rs.getByte("Wis"));
+			pc.addBaseStr(rs.getShort("Str"));
+			pc.addBaseCon(rs.getShort("Con"));
+			pc.addBaseDex(rs.getShort("Dex"));
+			pc.addBaseCha(rs.getShort("Cha"));
+			pc.addBaseInt(rs.getShort("Intel"));
+			pc.addBaseWis(rs.getShort("Wis"));
 			int status = rs.getInt("Status");
 			pc.setCurrentWeapon(status);
 			int classId = rs.getInt("Class");
